file_analysis/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
simhash==2.1.2
SQLAlchemy==2.0.30
asyncpg==0.29.0
pydantic-settings==2.9.1
